add `ShardReaderActor` to handle read flow
