+++ nsdb-cluster/src/main/scala/io/radicalbit/nsdb/cluster/actor/MetricsDataActor.scala
@@ -22,7 +22,7 @@ import java.util.concurrent.TimeUnit
 import akka.actor.{Actor, ActorLogging, ActorRef, PoisonPill, Props}
 import akka.pattern.{ask, pipe}
 import akka.util.Timeout
-import io.radicalbit.nsdb.actors.{ShardAccumulatorActor, ShardKey}
+import io.radicalbit.nsdb.actors.{ShardAccumulatorActor, ShardKey, ShardReaderActor}
 import io.radicalbit.nsdb.cluster.actor.MetricsDataActor._
 import io.radicalbit.nsdb.cluster.index.Location
 import io.radicalbit.nsdb.common.protocol.Bit
@@ -41,35 +41,38 @@ class MetricsDataActor(val basePath: String) extends Actor with ActorLogging {
 
   lazy val sharding: Boolean = context.system.settings.config.getBoolean("nsdb.sharding.enabled")
 
-  val childActors: mutable.Map[NamespaceKey, ActorRef] = mutable.Map.empty
+  val childActors: mutable.Map[NamespaceKey, (ActorRef, ActorRef)] = mutable.Map.empty
 
   /**
-    * Gets or creates child actor of class [[ShardAccumulatorActor]] to handle write requests
+    * Gets or creates reader child actor of class [[io.radicalbit.nsdb.actors.ShardReaderActor]] to handle read requests
     *
     * @param db database name
     * @param namespace namespace name
-    * @return [[ShardAccumulatorActor]] for selected database and namespace
+    * @return [[(ShardReaderActor, ShardAccumulatorActor)]] for selected database and namespace
     */
-  private def getOrCreateChild(db: String, namespace: String): ActorRef =
+  private def getOrCreateChildren(db: String, namespace: String): (ActorRef, ActorRef) =
     childActors.getOrElse(
       NamespaceKey(db, namespace), {
-        val child = context.actorOf(ShardAccumulatorActor.props(basePath, db, namespace),
-                                    s"shard-accumulator-service-$db-$namespace")
-        childActors += (NamespaceKey(db, namespace) -> child)
-        child
+        val reader =
+          context.actorOf(ShardReaderActor.props(basePath, db, namespace), s"shard-reader-$db-$namespace")
+
+        val accumulator = context.actorOf(ShardAccumulatorActor.props(basePath, db, namespace, reader),
+                                          s"shard-accumulator-$db-$namespace")
+        childActors += (NamespaceKey(db, namespace) -> (reader, accumulator))
+        (reader, accumulator)
       }
     )
 
   /**
-    * If exists, gets child for selected namespace and database.
-    * Use in case of read-only operations such as get metrics operations and ddl operations such as drop metrics.
+    * If exists, gets the reader for selected namespace and database.
+    * Use in case of read
     *
     * @param db database name
     * @param namespace namespace name
     * @return Option containing child actor of class [[ShardAccumulatorActor]]
     */
-  private def getChild(db: String, namespace: String): Option[ActorRef] =
-    childActors.get(NamespaceKey(db, namespace))
+  private def getReader(db: String, namespace: String): Option[ActorRef] =
+    childActors.get(NamespaceKey(db, namespace)).map(_._1)
 
   implicit val timeout: Timeout = Timeout(
     context.system.settings.config.getDuration("nsdb.namespace-data.timeout", TimeUnit.SECONDS),
@@ -87,9 +90,7 @@ class MetricsDataActor(val basePath: String) extends Actor with ActorLogging {
       })
       .foreach {
         case (db, namespace) =>
-          childActors += (NamespaceKey(db, namespace) -> context.actorOf(
-            ShardAccumulatorActor.props(basePath, db, namespace),
-            s"shard-accumulator-service-$db-$namespace"))
+          getOrCreateChildren(db, namespace)
       }
   }
 
@@ -101,12 +102,15 @@ class MetricsDataActor(val basePath: String) extends Actor with ActorLogging {
     case GetNamespaces(db) =>
       sender() ! NamespacesGot(db, childActors.keys.filter(_.db == db).map(_.namespace).toSet)
     case msg @ GetMetrics(db, namespace) =>
-      getChild(db, namespace) match {
+      getReader(db, namespace) match {
         case Some(child) => child forward msg
         case None        => sender() ! MetricsGot(db, namespace, Set.empty)
       }
     case DeleteNamespace(db, namespace) =>
-      val indexToRemove = getOrCreateChild(db, namespace)
+      val indexToRemove = getOrCreateChildren(db, namespace)._2
+
+      val c = childActors
+
       (indexToRemove ? DeleteAllMetrics(db, namespace))
         .map(_ => {
           indexToRemove ! PoisonPill
@@ -114,26 +118,29 @@ class MetricsDataActor(val basePath: String) extends Actor with ActorLogging {
           NamespaceDeleted(db, namespace)
         })
         .pipeTo(sender())
-    case msg @ DropMetric(db, namespace, metric) =>
-      getChild(db, namespace) match {
+    case msg @ DropMetric(db, namespace, _) =>
+      getOrCreateChildren(db, namespace)._2 forward msg
+    case msg @ GetCount(db, namespace, metric) =>
+      getReader(db, namespace) match {
         case Some(child) => child forward msg
-        case None        => sender() ! MetricDropped(db, namespace, metric)
+        case None        => sender() ! CountGot(db, namespace, metric, 0)
       }
-    case msg @ GetCount(db, namespace, _) =>
-      getOrCreateChild(db, namespace).forward(msg)
     case msg @ ExecuteSelectStatement(statement, _) =>
-      getOrCreateChild(statement.db, statement.namespace).forward(msg)
+      getReader(statement.db, statement.namespace) match {
+        case Some(child) => child forward msg
+        case None        => sender() ! SelectStatementExecuted(statement.db, statement.namespace, statement.metric, Seq.empty)
+      }
   }
 
   def shardBehaviour: Receive = {
     case AddRecordToLocation(db, namespace, bit, location) =>
-      getOrCreateChild(db, namespace).forward(
-        AddRecordToShard(db, namespace, ShardKey(location.metric, location.from, location.to), bit))
+      getOrCreateChildren(db, namespace)._2
+        .forward(AddRecordToShard(db, namespace, ShardKey(location.metric, location.from, location.to), bit))
     case DeleteRecordFromLocation(db, namespace, bit, location) =>
-      getOrCreateChild(db, namespace).forward(
-        DeleteRecordFromShard(db, namespace, ShardKey(location.metric, location.from, location.to), bit))
+      getOrCreateChildren(db, namespace)._2
+        .forward(DeleteRecordFromShard(db, namespace, ShardKey(location.metric, location.from, location.to), bit))
     case ExecuteDeleteStatementInternalInLocations(statement, schema, locations) =>
-      getOrCreateChild(statement.db, statement.namespace).forward(
+      getOrCreateChildren(statement.db, statement.namespace)._2.forward(
         ExecuteDeleteStatementInShards(statement, schema, locations.map(l => ShardKey(l.metric, l.from, l.to))))
   }
